pyproject.toml
[project]
name = "speakwith"
version = "0.1.0"
description = "Communication assistant for people who cannot speak but can read"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faster-whisper>=1.0.0",
    "openai>=1.0.0",
    "sounddevice>=0.5.0",
    "numpy>=1.26.0",
    "python-dotenv>=1.0.0",
    "aioconsole>=0.8.0",
    "rich>=13.0.0",
]

[project.scripts]
speakwith = "speakwith.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["speakwith"]
